requirements.txt
flask
pdf2image
pytesseract
python-docx
werkzeug
